# Replace JSONMarshal with MarshalText (and friends).