Would someone like to review my updates to switch to the more generic MarshalText/Binary routines from the JSONMarshal routines?