@jboverfelt 2016-02-13
Nice! I'll make a pull request to use these functions in the Scan() function once they're merged down.