@bmatsuo 2016-02-16
Looks OK to me. Perhaps unfortunate that the Unmarshal routines require conversion to a string (though number of allocations is unchanged). However, I do see the Unmarshal benchmark getting slower on my machine.

```
$ go test -test.bench=. -test.run=None -test.benchtime=2s -test.benchmem
```


```
benchmark                         old ns/op     new ns/op     delta
BenchmarkUUID_MarshalJSON-4       933           694           -25.62%
BenchmarkUUID_UnmarshalJSON-4     1194          1272          +6.53%
BenchmarkParse-4                  88.2          87.1          -1.25%
BenchmarkNew-4                    1793          1785          -0.45%
BenchmarkUUID_String-4            117           114           -2.56%
BenchmarkUUID_URN-4               122           119           -2.46%
BenchmarkUUID_Array-4             14.4          14.9          +3.47%
BenchmarkArray_UUID-4             5.65          5.80          +2.65%

benchmark                         old allocs     new allocs     delta
BenchmarkUUID_MarshalJSON-4       5              4              -20.00%
BenchmarkUUID_UnmarshalJSON-4     4              4              +0.00%
BenchmarkParse-4                  1              1              +0.00%
BenchmarkNew-4                    2              2              +0.00%
BenchmarkUUID_String-4            1              1              +0.00%
BenchmarkUUID_URN-4               1              1              +0.00%
BenchmarkUUID_Array-4             0              0              +0.00%
BenchmarkArray_UUID-4             0              0              +0.00%

benchmark                         old bytes     new bytes     delta
BenchmarkUUID_MarshalJSON-4       360           264           -26.67%
BenchmarkUUID_UnmarshalJSON-4     320           320           +0.00%
BenchmarkParse-4                  16            16            +0.00%
BenchmarkNew-4                    64            64            +0.00%
BenchmarkUUID_String-4            48            48            +0.00%
BenchmarkUUID_URN-4               48            48            +0.00%
BenchmarkUUID_Array-4             0             0             +0.00%
BenchmarkArray_UUID-4             0             0             +0.00%
```

By creating unexported versions of `Parse()` and `xtob()` that take `[]byte` arguments I was able to eliminate the unnecessary allocation and speed up the unmarshal benchmark a bit. But I didn't see it get faster than before.

```
benchmark                         old ns/op     new ns/op     delta
BenchmarkUUID_MarshalJSON-4       933           697           -25.29%
BenchmarkUUID_UnmarshalJSON-4     1194          1219          +2.09%
BenchmarkParse-4                  88.2          84.0          -4.76%
BenchmarkNew-4                    1793          1787          -0.33%
BenchmarkUUID_String-4            117           113           -3.42%
BenchmarkUUID_URN-4               122           120           -1.64%
BenchmarkUUID_Array-4             14.4          14.4          +0.00%
BenchmarkArray_UUID-4             5.65          5.61          -0.71%

benchmark                         old allocs     new allocs     delta
BenchmarkUUID_MarshalJSON-4       5              4              -20.00%
BenchmarkUUID_UnmarshalJSON-4     4              3              -25.00%
BenchmarkParse-4                  1              1              +0.00%
BenchmarkNew-4                    2              2              +0.00%
BenchmarkUUID_String-4            1              1              +0.00%
BenchmarkUUID_URN-4               1              1              +0.00%
BenchmarkUUID_Array-4             0              0              +0.00%
BenchmarkArray_UUID-4             0              0              +0.00%

benchmark                         old bytes     new bytes     delta
BenchmarkUUID_MarshalJSON-4       360           264           -26.67%
BenchmarkUUID_UnmarshalJSON-4     320           272           -15.00%
BenchmarkParse-4                  16            16            +0.00%
BenchmarkNew-4                    64            64            +0.00%
BenchmarkUUID_String-4            48            48            +0.00%
BenchmarkUUID_URN-4               48            48            +0.00%
BenchmarkUUID_Array-4             0             0             +0.00%
BenchmarkArray_UUID-4             0             0             +0.00%
```