@pborman 2017-01-12
Thanks for the prod.  Is there a reason?  No, probably not, other than I was hoping github.com/google/uuid would come to replace this one.  It has a slightly different API (UUID is an array rather than a slice) so it is not drop in.